test : article-dao done!

+++ back_end/org.gobeshona/src/main/java/org/gobeshona/articles/Article_Dao_I.java
@@ -4,11 +4,11 @@ import org.gobeshona.articles.Articles;
 
 /**
  *
- * @author Sajib
+ * @author abedeen
  */
 public interface Article_Dao_I 
 {
     public Articles post_Article(Articles articles);
-    public void delete_Article(Long articleId);
+    public void delete_Article(Integer articleId);
     public void edit_Article(Articles articles);    
 }//ArticlePersistence_I
